+++ pytorch_lightning/tuner/auto_gpu_select.py
@@ -22,7 +22,8 @@ def pick_multiple_gpus(nb: int) -> List[int]:
     """
     Raises:
         MisconfigurationException:
-            If ``gpus`` is set to 0, when ``auto_select_gpus=True``.
+            If ``gpus`` or ``devices`` is set to 0, when ``auto_select_gpus=True``, or when the requested number is
+            higher than the number of GPUs available on the machine.
     """
     if nb == 0:
         raise MisconfigurationException(
@@ -30,7 +31,10 @@ def pick_multiple_gpus(nb: int) -> List[int]:
             " Please select a valid number of GPU resources when using auto_select_gpus."
         )
 
-    nb = torch.cuda.device_count() if nb == -1 else nb
+    num_gpus = torch.cuda.device_count()
+    if nb > num_gpus:
+        raise MisconfigurationException(f"You requested {nb} GPUs but your machine only has {num_gpus} GPUs.")
+    nb = num_gpus if nb == -1 else nb
 
     picked: List[int] = []
     for _ in range(nb):