+++ tests/deprecated_api/test_remove_1-7.py
@@ -20,10 +20,8 @@ from unittest.mock import Mock
 import pytest
 import torch
 
-import pytorch_lightning
 from pytorch_lightning import Callback, LightningDataModule, Trainer
 from pytorch_lightning.callbacks.lr_monitor import LearningRateMonitor
-from pytorch_lightning.callbacks.xla_stats_monitor import XLAStatsMonitor
 from pytorch_lightning.loggers import LoggerCollection, TestTubeLogger
 from pytorch_lightning.overrides.distributed import IndexBatchSamplerWrapper
 from pytorch_lightning.plugins.environments import (
@@ -34,18 +32,11 @@ from pytorch_lightning.plugins.environments import (
     TorchElasticEnvironment,
 )
 from pytorch_lightning.strategies import SingleDeviceStrategy
-from pytorch_lightning.utilities.exceptions import MisconfigurationException
 from tests.deprecated_api import _soft_unimport_module
 from tests.helpers import BoringModel
 from tests.helpers.datamodules import MNISTDataModule
 from tests.loggers.test_logger import CustomLogger
 from tests.plugins.environments.test_lsf_environment import _make_rankfile
-
-
-def test_v1_7_0_moved_model_summary_and_layer_summary(tmpdir):
-    _soft_unimport_module("pytorch_lightning.core.memory")
-    with pytest.deprecated_call(match="to `pytorch_lightning.utilities.model_summary` since v1.5"):
-        from pytorch_lightning.core.memory import LayerSummary, ModelSummary  # noqa: F401
 
 
 def test_v1_7_0_datamodule_transform_properties(tmpdir):
@@ -289,11 +280,6 @@ def test_v1_7_0_deprecate_parameter_validation():
         from pytorch_lightning.core.decorators import parameter_validation  # noqa: F401
 
 
-def test_v1_7_0_passing_strategy_to_plugins_flag():
-    with pytest.deprecated_call(match="has been deprecated in v1.5 and will be removed in v1.7."):
-        Trainer(plugins="ddp_spawn")
-
-
 def test_v1_7_0_weights_summary_trainer(tmpdir):
     with pytest.deprecated_call(match=r"Setting `Trainer\(weights_summary=full\)` is deprecated in v1.5"):
         t = Trainer(weights_summary="full")
@@ -313,12 +299,6 @@ def test_v1_7_0_deprecated_slurm_job_id():
     trainer = Trainer()
     with pytest.deprecated_call(match="Method `slurm_job_id` is deprecated in v1.6.0 and will be removed in v1.7.0."):
         trainer.slurm_job_id
-
-
-def test_v1_7_0_deprecate_xla_stats_monitor(monkeypatch):
-    monkeypatch.setattr(pytorch_lightning.callbacks.xla_stats_monitor, "_TPU_AVAILABLE", True)
-    with pytest.deprecated_call(match="The `XLAStatsMonitor` callback was deprecated in v1.5"):
-        _ = XLAStatsMonitor()
 
 
 def test_v1_7_0_deprecated_max_steps_none(tmpdir):
@@ -433,17 +413,3 @@ def test_v1_7_0_post_dispatch_hook():
 
     with pytest.deprecated_call(match=escape("`CustomPlugin.post_dispatch()` has been deprecated in v1.6")):
         CustomPlugin(torch.device("cpu"))
-
-
-def test_xla_stats_monitor_tpu_not_used(monkeypatch):
-    monkeypatch.setattr(pytorch_lightning.callbacks.xla_stats_monitor, "_TPU_AVAILABLE", True)
-    with pytest.deprecated_call(match="The `XLAStatsMonitor` callback was deprecated in v1.5"):
-        xla_stats = XLAStatsMonitor()
-
-    trainer = Trainer(accelerator="cpu", callbacks=[xla_stats])
-    model = BoringModel()
-    with pytest.raises(
-        MisconfigurationException,
-        match="You are using XLAStatsMonitor but are not running on TPU. The accelerator is set to CPUAccelerator.",
-    ):
-        trainer.fit(model)